Integrated cmark's parser and HTML converter to our project. Implemented via the markdown_converter.h/.cc files. Added unit testing. Code coverage for markdown_converter is at 100%. The total coverage for lines is 92.1%. The total coverage for functions is 94.6%. Had to update the base.Dockerfile to include the libcmark library for cmark as well as update the CMake list to properly link the library.
Update: Addressed the 3 comments. Updated OPT_DEFAULT to OPT_SAFE to make sure we always strip raw HTML. Wrapped the free(html) in a nullptr check. Changed the normalize newlines in the test to skipping any \r for security. With these changes, potential nullptr is checked for and the security concerns of raw HTML and sneaking in a lone \r is averted. Let me know what you think!

Change-Id: I93eb7e2a36bfc1ce49310f4128025423cdf8fad1